requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
openai==1.30.1
anthropic==0.28.0
slack-sdk==3.27.2
streamlit==1.35.0
plotly==5.22.0
schedule==1.2.2
python-dotenv==1.0.1
pydantic==2.7.1
rich==13.7.1
httpx==0.27.0
pytest==8.2.0
